refactor(ipc): D4 — split server.ts into namespace directory
Replaces the 1239-LOC packages/gateway/src/ipc/server.ts with a
directory of concern-focused sibling files:

- index.ts            — re-export shim (3 public symbols)
- rpc-error.ts        — RpcMethodError class (universal error type)
- options.ts          — CreateIpcServerOptions type + BunSessionData
- context.ts          — ServerCtx interface + 6 skip-symbol exports
- vault-dispatch.ts   — vault gate + dispatchVaultGated (public) +
                        rpcVaultOrMethodNotFound
- socket-listeners.ts — Win32 + Bun unix listener wiring + helpers
- inline-handlers.ts  — direct method handlers (gateway.ping,
                        index.searchRanked, consent.respond,
                        audit.list, agent.invoke, engine.askStream,
                        workflow.run + helpers)
- dispatchers.ts      — all 14 tryDispatchXxxRpc free functions
                        taking ServerCtx (largest at 526 LOC)
- server.ts           — createIpcServer factory + closure state +
                        thin dispatchMethod + IPCServer object

Mostly mechanical move; zero behavioral change. Per-namespace
dispatchers and inline handlers become free functions accepting a
constructor-bound ServerCtx (mirrors MeshSpawnContext pattern from
PR #163).

agentInvokeHandler / workflowRunHandler are exposed as getters on
ServerCtx (ctx.getAgentInvokeHandler() etc.) so the public
setAgentInvokeHandler / setWorkflowRunHandler API still affects
future dispatched calls. Each handler-using function captures the
getter result ONCE at the same point the original closure read it
(per plan-review §2.2 — re-reading the getter inside an async block
could see a different handler if the setter fires mid-stream).

The largest resulting file is dispatchers.ts at 526 LOC, well under
the 800 D4 threshold.

Four consumer files (ipc/index.ts, engine-ask-stream.test.ts,
ipc.test.ts, server-vault-gated.test.ts) update their import path
from "./server.ts" to "./server/index.ts".

Spec: docs/superpowers/specs/2026-05-02-d4-server-split-design.md
Plan: docs/superpowers/plans/2026-05-02-d4-server-split.md
