+++ packages/gateway/src/ipc/index.ts
@@ -6,7 +6,7 @@
 
 export type { AgentInvokeContext, AgentInvokeHandler } from "./agent-invoke.ts";
 export { type ConsentCoordinator, ConsentDisconnectedError } from "./consent.ts";
-export type { CreateIpcServerOptions } from "./server.ts";
-export { createIpcServer } from "./server.ts";
+export type { CreateIpcServerOptions } from "./server/index.ts";
+export { createIpcServer } from "./server/index.ts";
 export type { IPCServer } from "./types.ts";
 export type { WorkflowRunContext, WorkflowRunHandler } from "./workflow-invoke.ts";
