+++ packages/gateway/src/ipc/server/index.ts
@@ -1,0 +1,3 @@
+export type { CreateIpcServerOptions } from "./options.ts";
+export { createIpcServer } from "./server.ts";
+export { dispatchVaultGated } from "./vault-dispatch.ts";
